refactor: refactor employee table to move utility functions to utility folder
