+++ src/components/filters/EmployeeFilters.tsx
@@ -61,7 +61,7 @@ const EmployeeFilters: React.FC<EmployeeFiltersProps> = ({
 
       {/* Advanced Filters */}
       {showFilters && (
-        <div className="bg-input border border-border rounded-lg p-4">
+        <div className="bg-popover border border-border rounded-lg p-4">
           <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-4">
             <div className="space-y-2">
               <Label htmlFor="department-filter">Department</Label>
