+++ src/components/tables/EmployeeTable.tsx
@@ -6,12 +6,9 @@ import {
   getSortedRowModel,
   getFilteredRowModel,
   flexRender,
-  type ColumnDef,
-  type Row,
 } from '@tanstack/react-table';
 import { ArrowUpDown, Users } from 'lucide-react';
 import {
-  Checkbox,
   Button,
   Table,
   TableHeader,
@@ -20,12 +17,10 @@ import {
   TableHead,
   TableCell,
 } from '@/components/ui';
-import { Avatar } from '@/components/common/Avatar';
-import { ActionsDropdown } from './ActionsDropdown';
-import { getAvatarUrl, formatDate } from '@/utils/formatting';
-import type { Employee } from '@/types';
 import type { EmployeeTableProps } from './types';
 import { EMPLOYEES_PER_PAGE } from '@/constants/data';
+import { globalEmployeeFilter } from '@/utils/filters';
+import { getEmployeeColumns } from './employeeColumns';
 
 export const EmployeeTable: React.FC<EmployeeTableProps> = ({
   employees,
@@ -40,117 +35,9 @@ export const EmployeeTable: React.FC<EmployeeTableProps> = ({
   onEdit,
   onDelete,
 }) => {
-  // Custom global filter function
-  const globalFilterFn = (row: Row<Employee>, _columnId: string, value: string) => {
-    const employee = row.original as Employee;
-    const searchValue = value.toLowerCase();
-
-    const firstName = (employee.firstName || '').toLowerCase();
-    const lastName = (employee.lastName || '').toLowerCase();
-    const email = (employee.email || '').toLowerCase();
-    const fullName = `${firstName} ${lastName}`;
-
-    return (
-      firstName.includes(searchValue) ||
-      lastName.includes(searchValue) ||
-      fullName.includes(searchValue) ||
-      email.includes(searchValue)
-    );
-  };
-
   // Define columns for TanStack Table
-  const columns = useMemo<ColumnDef<Employee>[]>(
-    () => [
-      {
-        id: 'select',
-        header: ({ table }) => (
-          <Checkbox
-            checked={table.getIsAllPageRowsSelected()}
-            onCheckedChange={value => table.toggleAllPageRowsSelected(!!value)}
-            aria-label="Select all"
-          />
-        ),
-        cell: ({ row }) => (
-          <Checkbox
-            checked={row.getIsSelected()}
-            onCheckedChange={value => row.toggleSelected(!!value)}
-            aria-label="Select row"
-          />
-        ),
-        enableSorting: false,
-        enableHiding: false,
-        enableGlobalFilter: false,
-      },
-      {
-        accessorKey: 'name',
-        header: 'Name',
-        cell: ({ row }) => {
-          const employee = row.original;
-          return (
-            <div className="flex items-center gap-3">
-              <Avatar
-                name={`${employee.firstName} ${employee.lastName}`}
-                imageUrl={
-                  employee.profilePhoto || getAvatarUrl(employee.firstName, employee.lastName)
-                }
-                size="sm"
-              />
-              <div>
-                <div className="font-medium text-primary">
-                  {employee.firstName} {employee.lastName}
-                </div>
-                <div className="text-sm text-primary">{(employee.email || '').toLowerCase()}</div>
-              </div>
-            </div>
-          );
-        },
-        enableSorting: true,
-        sortingFn: (rowA, rowB) => {
-          const nameA = `${rowA.original.firstName} ${rowA.original.lastName}`;
-          const nameB = `${rowB.original.firstName} ${rowB.original.lastName}`;
-          return nameA.localeCompare(nameB);
-        },
-        enableGlobalFilter: true,
-        accessorFn: row => `${row.firstName} ${row.lastName} ${row.email}`,
-      },
-      {
-        accessorKey: 'hireDate',
-        header: 'Hire Date',
-        cell: ({ row }) => (
-          <div className="text-sm text-primary">{formatDate(row.original.hireDate)}</div>
-        ),
-        enableSorting: true,
-        enableGlobalFilter: false,
-      },
-      {
-        accessorKey: 'jobTitle',
-        header: 'Job Title',
-        cell: ({ row }) => <div className="text-sm text-primary">{row.original.jobTitle}</div>,
-        enableSorting: true,
-        enableGlobalFilter: false,
-      },
-      {
-        accessorKey: 'contractType',
-        header: 'Employment Type',
-        cell: ({ row }) => <div className="text-sm text-primary">{row.original.contractType}</div>,
-        enableSorting: true,
-        enableGlobalFilter: false,
-      },
-      {
-        id: 'actions',
-        header: '',
-        cell: ({ row }) => (
-          <ActionsDropdown
-            employee={row.original}
-            onView={onView}
-            onEdit={onEdit}
-            onDelete={onDelete}
-          />
-        ),
-        enableSorting: false,
-        enableGlobalFilter: false,
-      },
-    ],
+  const columns = useMemo(
+    () => getEmployeeColumns(onView, onEdit, onDelete),
     [onView, onEdit, onDelete]
   );
 
@@ -172,7 +59,7 @@ export const EmployeeTable: React.FC<EmployeeTableProps> = ({
     getFilteredRowModel: getFilteredRowModel(),
     getPaginationRowModel: getPaginationRowModel(),
     getSortedRowModel: getSortedRowModel(),
-    globalFilterFn: globalFilterFn,
+    globalFilterFn: globalEmployeeFilter,
     initialState: {
       pagination: {
         pageSize: EMPLOYEES_PER_PAGE,
